Refactor: builder 패턴 변경, properties 추가, 패키지 구조 변경
